feat: adds copy button to scim url and bearer token (#10)
During the self service scim setup we are missing copy buttons for the SCIM URL and the generated bearer token. Now we have extracted the copy value logic we already use in the SAML setup and also use it in the SCIM setup.
